backend/stock-analytics-service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
motor==3.5.1
pydantic==2.9.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
google-generativeai==0.7.2
vnstock>=0.2.3
pandas==2.2.3
plotly==5.24.1